feat: swagger doc for direct upload refactoring: moved create project developer api to users module
