+++ backend/app/controllers/api/v1/project_developers_controller.rb
@@ -3,9 +3,6 @@ module API
     class ProjectDevelopersController < BaseController
       include API::Pagination
 
-      skip_before_action :require_json!, only: :create
-      before_action :authenticate_user!, only: :create
-
       def index
         project_developers = ProjectDeveloper.all.includes(:account)
         pagy_object, project_developers = pagy(project_developers, page: current_page, items: per_page)
@@ -26,17 +23,6 @@ module API
         ).serializable_hash
       end
 
-      def create
-        current_user.with_lock do
-          raise API::UnprocessableEntityError, I18n.t("errors.messages.user_multiple_accounts") if current_user.account_id.present?
-
-          account = Account.create! account_params
-          current_user.update! account: account, role: :project_developer
-          project_developer = ProjectDeveloper.create! project_developer_params.merge account: account
-          render json: ProjectDeveloperSerializer.new(project_developer).serializable_hash
-        end
-      end
-
       private
 
       def fetch_project_developer
@@ -44,17 +30,6 @@ module API
 
         account = Account.friendly.find(params[:id])
         ProjectDeveloper.find_by!(account_id: account.id)
-      end
-
-      def account_params
-        params.fetch(:project_developer_params, params)
-          .permit :language, :picture, :name, :website, :linkedin, :facebook, :twitter, :instagram, :about
-      end
-
-      def project_developer_params
-        params.fetch(:project_developer_params, params)
-          .permit :language, :project_developer_type, :entity_legal_registration_number, :mission,
-            categories: [], impacts: [], location_ids: []
       end
     end
   end
