+++ backend/spec/fixtures/snapshots/api/v1/get-direct-upload.json
@@ -1,0 +1,20 @@
+{
+  "id": "0d078da9-f5c7-4c2d-be24-45fdb49d6589",
+  "key": "ldqd8f8r83zapvsevn43hb1yl39x",
+  "filename": "test.jpg",
+  "content_type": "image/jpeg",
+  "metadata": {
+  },
+  "service_name": "test",
+  "byte_size": 32326,
+  "checksum": "QYeLAwqIj9HrwITqtTYaEw==",
+  "created_at": "2022-03-25T10:52:54.826Z",
+  "signed_id": "eyJfcmFpbHMiOnsibWVzc2FnZSI6IkJBaEpJaWt3WkRBM09HUmhPUzFtTldNM0xUUmpNbVF0WW1VeU5DMDBOV1prWWpRNVpEWTFPRGtHT2daRlZBPT0iLCJleHAiOm51bGwsInB1ciI6ImJsb2JfaWQifX0=--c7f6a95a696eefe8eb3689bf8e585069819bddbb",
+  "attachable_sgid": "BAh7CEkiCGdpZAY6BkVUSSJWZ2lkOi8vYmFja2VuZC9BY3RpdmVTdG9yYWdlOjpCbG9iLzBkMDc4ZGE5LWY1YzctNGMyZC1iZTI0LTQ1ZmRiNDlkNjU4OT9leHBpcmVzX2luBjsAVEkiDHB1cnBvc2UGOwBUSSIPYXR0YWNoYWJsZQY7AFRJIg9leHBpcmVzX2F0BjsAVDA=--bf2bf73b10c399645220c833dab610acf70ffd92",
+  "direct_upload": {
+    "url": "http://www.example.com/rails/active_storage/disk/eyJfcmFpbHMiOnsibWVzc2FnZSI6IkJBaDdDam9JYTJWNVNTSWhiR1J4WkRobU9ISTRNM3BoY0haelpYWnVORE5vWWpGNWJETTVlQVk2QmtWVU9oRmpiMjUwWlc1MFgzUjVjR1ZKSWc5cGJXRm5aUzlxY0dWbkJqc0dWRG9UWTI5dWRHVnVkRjlzWlc1bmRHaHBBa1orT2cxamFHVmphM04xYlVraUhWRlpaVXhCZDNGSmFqbEljbmRKVkhGMFZGbGhSWGM5UFFZN0JsUTZFWE5sY25acFkyVmZibUZ0WlRvSmRHVnpkQT09IiwiZXhwIjoiMjAyMi0wMy0yNVQxMDo1Nzo1NC44NDRaIiwicHVyIjoiYmxvYl90b2tlbiJ9fQ==--2373415a32ee768035f04a7316cce3eb0b25dd19",
+    "headers": {
+      "Content-Type": "image/jpeg"
+    }
+  }
+}
